+++ lib/kward/prompt_interface.rb
@@ -25,6 +25,7 @@ require_relative "prompt_interface/overlay_renderer"
 require_relative "prompt_interface/editor/renderer"
 require_relative "prompt_interface/editor/syntax_highlighter"
 require_relative "prompt_interface/editor/auto_close_pairs"
+require_relative "prompt_interface/editor/endwise"
 require_relative "prompt_interface/editor/auto_indent"
 require_relative "prompt_interface/composer_renderer"
 require_relative "prompt_interface/composer_controller"
@@ -72,6 +73,7 @@ module Kward
     include EditorRenderer
     include EditorSyntaxHighlighter
     include EditorAutoClosePairs
+    include EditorEndwise
     include EditorAutoIndent
     include ComposerRenderer
     include ComposerController